ebuild-bumper.sh: Modified clean function
1. Changed sort of list of ebuilds to clean
2. Added code to skip removing any ebuilds if only a 9999 version and 1
   other exist

+++ ebuild-bumper.sh
@@ -160,12 +160,15 @@ clean() {
 		cd ${REPO}/${CAT}/${my_pn}/
 
 		# Find all ebuilds, sorted, in array
-		local EBUILDS=( $( ls *ebuild | sort -Vr ) )
+		local EBUILDS=( $( ls *ebuild | LC_COLLATE=C sort -t. -k3 ) )
 
 		[[ ${#EBUILDS[@]} -le 1 ]] && continue
 
 		local start=1
-		[[ ${EBUILDS[@]} == *"9999"* ]] && let start=2
+		if [[ ${EBUILDS[@]} == *"9999"* ]]; then
+			[[ ${#EBUILDS[@]} -eq 2 ]] && continue
+			let start=2
+		fi
 
 		for ebuild in ${EBUILDS[@]:${start}}; do
 			git rm ${ebuild}
